Add More Conditions To processOperator Function
On operator click, a boolean is checked to determine if
on the second number yet, if not, assign the inputted number
to numA and move it up to upper line with latest clicked
operator attached.
The main display is then cleared, ready for new input to determine
second number.
A boolean, isUpperLineActive, was created. If true upon operator
click, it simply changes the latest operator and this is represented
on the upper line, while the main display and munA remain unchanged